cs + loading check
